refactor(app): extract edit image input hook
